Comment the js.
